requirements.txt
alembic==1.11.1
blinker==1.6.2
cffi==1.15.1
click==8.1.4
colorama==0.4.6
cryptography==41.0.1
Flask==2.3.2
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
packaging==23.1
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
PyMySQL==1.1.0
SQLAlchemy==2.0.18
typing-extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
